DuplicateEntryException, InexistentEntryException, GameAlreadySettledException, GameNotSettledException : The empty constructor now calls the superconstructor by passing it an empty string, to avoid null pointer exceptions when using the getMessage() method.
The Game and Slot classes has been refactored to comply to the API
model. The same applies for RoundTest, SeasonTest and TournamentTest .

The Round and Season have been refactored to comply to PMD directives.

Test classes have been written for all the exception classes as well as
the Game class.

At this point, coverage for the model package is at 100% and there are
no PMD violations.